Fix simple goal when a pom is missing Submitted By: Lester Ecarma
git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk/maven-assembly-plugin@421659 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/maven/plugin/assembly/repository/DefaultRepositoryAssembler.java
@@ -34,7 +34,6 @@ import java.util.Set;
 import java.util.TimeZone;
 
 import org.apache.maven.artifact.Artifact;
-import org.apache.maven.artifact.ArtifactUtils;
 import org.apache.maven.artifact.factory.ArtifactFactory;
 import org.apache.maven.artifact.metadata.ArtifactMetadataSource;
 import org.apache.maven.artifact.repository.ArtifactRepository;
@@ -225,6 +224,12 @@ public class DefaultRepositoryAssembler
                             setAlignment( a );
 
                             File sourceFile = new File( localRepository.getBasedir(), localRepository.pathOf( a ) );
+
+                            if( !sourceFile.exists() )
+                            {
+                                break;
+                            }
+
                             targetFile = new File( targetRepository.getBasedir(), targetRepository.pathOf( a ) );
 
                             FileUtils.copyFile( sourceFile, targetFile );
